Created Status Effect Display

+++ Assets/Scripts/User Interface/StatusBar.cs
@@ -5,13 +5,13 @@ public class StatusBar : MonoBehaviour
 {
     public Slider slider;
 
-    public void SetMaxValue(int val)
+    public void SetMaxValue(float val)
     {
         slider.maxValue = val;
         slider.value = val;
     }
 
-    public void SetValue(int val)
+    public void SetValue(float val)
     {
         slider.value = val;
     }
